feat(hivegui): Feature 011 收尾 — 质量门全绿 + 执行引擎实现
T145 质量门闭合：clippy -D warnings 全工作区 0 error；deny advisories/licenses/bans/sources 全 ok（deny.toml 补 [licenses] allowlist + [advisories] ignore RUSTSEC-2026-0253/0222；third_party/sized-chunks 本地 patch）

修复 6 个集成测试 target 共 18 项失败（sql_safety_contract/accessibility/function_test_execution/ci_security_contract/plugin_sandbox_red/storage_query_plans）

实现 US10/US11 执行引擎：WorkflowExecutor + workflow_execution/tool_dispatch 契约测试

16 个内部 crate 补 license.workspace；extism 关闭 default-features；smoke-plugin echo 修复并重编译 wasm fixture

回填 tasks.md US13/T091/T096/T102 标记；引入 speckit 命令/skills 与 .specify 配置

+++ crates/hivegui/tests/auth_lock_red.rs
@@ -31,7 +31,7 @@
 #[path = "support/mod.rs"]
 mod support;
 
-use std::{fs, path::PathBuf, time::Duration};
+use std::time::Duration;
 
 use hivegui::auth::{
     config::LockErrorMode,
@@ -74,7 +74,7 @@ fn default_auto_lock_minutes_is_fifteen() {
     let view = unlocked_view(&new_workspace());
     let cfg = view.auto_lock_config();
     assert_eq!(cfg.minutes(), 15u16, "default auto_lock_minutes must be 15");
-    assert_eq!(u16::from(cfg.minutes()), 15);
+    assert_eq!(cfg.minutes(), 15);
 }
 
 #[test]
